+++ CHANGELOG.md
@@ -1,6 +1,17 @@
-# CHANGE LOG
+# CHANGELOG
 
-## version 0.0.1
+#### version v1.0.1
+- 时间：2018-12-20
+```
+1. OnRoomDetail 参数修改，添加组队信息 BrigadeList
+2. 添加pushHandler主动接口 GetOffLineCacheData 断线重连获取断线期间的数据
+3. SetFrameSyncRate 接口添加 cacheFrameMs 参数，并改为结构体
+4. OnSetFrameSyncRate 接口添加 cacheFrameMs 参数
+5. FrameDataList 改为 MsFrameDataList
+6. 修复断线重连奔溃问题
+```
+
+#### version 0.0.1
 ```
 初次提交，只能本地调试
 ```